add AbstractStructSchema for Un-AvroBased StructSchema
