+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/schema/AvroSchema.java
@@ -40,7 +40,7 @@ import java.util.Map;
  * An AVRO schema implementation.
  */
 @Slf4j
-public class AvroSchema<T> extends AvroBasedStructSchema<T> {
+public class AvroSchema<T> extends StructSchema<T> {
     private static final Logger LOG = LoggerFactory.getLogger(AvroSchema.class);
 
     private ClassLoader pojoClassLoader;
